refactor: rename more sidebar folders to sideBar
